Add session ops utility with some facilitator functions
It is added as a main module since the utilities folder is reserved
only for utilities used by the core aexpect code or aexpect client.

The only detected linter problem was in an import done solely to
satisfy sphinx documentation links which can be reconsidered in
the future.

Another small change from the original pull request is in importing
python 2 compatible shell escaped strings which will hopefully no
longer be necessary in the future.

Signed-off-by: Plamen Dimitrov <pdimitrov@pevogam.com>